fixed get scores by id function and get workouts by id to return all user elements
