+++ server/controller/scoreController.js
@@ -2,7 +2,6 @@ const db = require("../db/db");
 const { createLog } = require("../support/createLog");
 
 const getAllScores = async (req, res) => {
-
   try {
     const result = await db("scores").select("*");
 
@@ -16,8 +15,7 @@ const getAllScores = async (req, res) => {
       });
     }
     res.status(200).send(result);
-
-  } catch(err) {
+  } catch (err) {
     createLog({
       method: "GET",
       action: "FETCH_SCORES",
@@ -27,7 +25,32 @@ const getAllScores = async (req, res) => {
     });
     res.status(500).send({ message: err });
   }
-
 };
 
-module.exports = { getAllScores };
+const getScoresById = async (req, res) => {
+  const { id } = req.params;
+  try {
+    const result = await db("scores").select("*").where("user_id", id);
+    if (result) {
+      createLog({
+        method: "GET",
+        action: "FETCH_USER_SCORES",
+        status_code: 200,
+        user_id: result.id,
+        metadata: JSON.stringify(result),
+      });
+    }
+    res.status(200).send(result);
+  } catch (err) {
+    createLog({
+      method: "GET",
+      action: "FETCH_USER_SCORES",
+      status_code: 500,
+      user_id: id,
+      metadata: err,
+    });
+    res.status(500).send({ message: err });
+  }
+};
+
+module.exports = { getAllScores, getScoresById };
