+++ server/routes/users.js
@@ -7,7 +7,6 @@ const {
   getWorkoutsById,
   getAllEvents,
   getEventsById,
-  getScoresById,
 } = require("../controller/userController");
 const router = express();
 
@@ -18,6 +17,5 @@ router.get("/user_goals/id/:id", getGoalsById);
 router.get("/user_workouts/id/:id", getWorkoutsById);
 router.get("/user_events/", getAllEvents);
 router.get("/user_events/id/:id", getEventsById);
-router.get("/scores/id/:id", getScoresById);
 
 module.exports = router;
